+++ src/backend/funmooc/gimporter/import_scripts/helpers.py
@@ -1,5 +1,6 @@
 import io
 import urllib
+from datetime import datetime
 
 from django.conf import settings
 from django.core.files import File
@@ -7,6 +8,7 @@ from django.core.files import File
 from cms.api import add_plugin, create_page
 from cms.models import Page
 from filer.models.imagemodels import Image
+from pytz import timezone
 from richie.apps.courses.defaults import PAGES_INFO
 
 
@@ -87,3 +89,29 @@ def create_or_update_single_plugin(
         plugin_type.model.objects.filter(id=plugin.id).update(
             plugin_type=plugin_type.__name__, **kwargs
         )
+
+
+def parse_date(date_string):
+    """
+    Parse a date string of the form "dd.mm.yy" to a timezone aware datetime assuming it is
+    Paris time
+    """
+    return (
+        timezone("Europe/Paris").localize(datetime.strptime(date_string, "%d.%m.%y"))
+        if date_string
+        else None
+    )
+
+
+def parse_datetime(datetime_string):
+    """
+    Parse a datetime string of the form "yyyy-mm-dd HH:MM:SS" and make it timezone aware
+    assuming it is Paris time.
+    """
+    return (
+        timezone("Europe/Paris").localize(
+            datetime.strptime(datetime_string, "%Y-%m-%d %H:%M:%S")
+        )
+        if datetime_string
+        else None
+    )